requirements.txt
django==5.2.1
django-ninja==1.4.1
django-ninja-jwt==5.3.7
pytest==8.3.5
pytest-django==4.11.1 